+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Modified
 - experience block for Daggerheart
+- homebrews creating/editing for Daggerheart
 
 ## [0.2.10] - 2025-07-28
 ### Added